flask-cloud-run/requirements.txt
#requirements.txt
Flask==2.3.2
google-cloud-storage==2.4.0
